# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peminjaman Service API",
        "description": "API Service Peminjaman Buku E-Library — IAE BBK2HAB3 | NIM: 102022400314",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Docker Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Ambil semua daftar peminjaman",
                "description": "Mengembalikan semua data peminjaman buku yang tersimpan di sistem.",
                "operationId": "7b7abc4c34a3af61f8848cb61741fd0e",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Loan"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Peminjaman-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Buat peminjaman buku baru",
                "description": "Membuat data peminjaman baru. Member mengajukan peminjaman akses E-book.",
                "operationId": "8018f8386b85532378c494c7b648e6e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "book_id",
                                    "book_title",
                                    "member_name",
                                    "loan_date",
                                    "due_date"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "string",
                                        "example": "MBR-001"
                                    },
                                    "book_id": {
                                        "type": "string",
                                        "example": "BOOK-123"
                                    },
                                    "book_title": {
                                        "type": "string",
                                        "example": "Clean Code"
                                    },
                                    "member_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-29"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Peminjaman e-book via platform"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peminjaman berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Loan created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Peminjaman-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/loans/{id}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Ambil detail peminjaman berdasarkan ID",
                "description": "Mengembalikan detail data satu peminjaman berdasarkan ID yang diberikan.",
                "operationId": "9330a8c8074b0405a752da7873822cbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID peminjaman",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Peminjaman-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Loan not found"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Loan": {
                "title": "Loan",
                "description": "Model data peminjaman buku E-Library",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "member_id": {
                        "type": "string",
                        "example": "MBR-001"
                    },
                    "book_id": {
                        "type": "string",
                        "example": "BOOK-123"
                    },
                    "book_title": {
                        "type": "string",
                        "example": "Clean Code"
                    },
                    "member_name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "loan_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-29"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": null,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "returned",
                            "overdue"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Peminjaman via platform",
                        "nullable": true
                    },
                    "audit_receipt": {
                        "type": "string",
                        "example": "IAE-LOG-2026-F349DF14",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key untuk autentikasi. Gunakan NIM: 102022400314",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Loans",
            "description": "Loans"
        }
    ]
}